requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.4.0
apimatic-core==0.2.6 
apimatic-core-interfaces==0.1.4 
apimatic-requests-client-adapter==0.1.6 
charset-normalizer==3.3.2 
deprecation==2.1.0 
enum34==1.1.10 
jsonpickle==3.0.2 
jsonpointer==2.4 
requests==2.31.0 
setuptools==66.0.0 
squareup==34.0.1.20240118